+++ packages/instant-meilisearch/src/adapter/search-request-adapter/search-params-adapter.ts
@@ -238,6 +238,18 @@ export function MeiliParamsCreator(searchContext: SearchContext) {
         meiliSearchParams.hybrid = value
       }
     },
+    addDistinct() {
+      const value = overrideParams?.distinct
+      if (value !== undefined) {
+        meiliSearchParams.distinct = value
+      }
+    },
+    addRankingScoreThreshold() {
+      const value = overrideParams?.rankingScoreThreshold
+      if (value !== undefined) {
+        meiliSearchParams.rankingScoreThreshold = value
+      }
+    },
   }
 }
 
@@ -270,6 +282,8 @@ export function adaptSearchParams(
   meilisearchParams.addShowRankingScore()
   meilisearchParams.addAttributesToSearchOn()
   meilisearchParams.addHybridSearch()
+  meilisearchParams.addDistinct()
+  meilisearchParams.addRankingScoreThreshold()
 
   return meilisearchParams.getParams()
 }